+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Added
 - Algorithms:
     - Random fanout substitution strategy
     - Maximum-effort mode in `gold` utilizing random fanout substitution strategies and random topological orderings to design high-quality layouts
+    - Flag in `gold` to enforce NOT gates to be routed non-bending only
 
 
 v0.6.11 - 2025-04-23